+++ src/index.js
@@ -1,13 +1,16 @@
 const express = require("express");
-const { PrismaClient } = require("@prisma/client");
 const app = express();
 const port = 4000;
 const cors = require("cors");
-const jwt = require("jsonwebtoken");
-const prisma = new PrismaClient();
 const swaggerUi = require("swagger-ui-express");
 const swaggerJsDoc = require("swagger-jsdoc");
 require("dotenv").config();
+const {
+  handleUserSignUp,
+  handleUserLogin,
+  handleUserLogout,
+} = require("./controllers/user.controller.js");
+const sever_ip = process.env.IP;
 
 const options = {
   swaggerDefinition: {
@@ -19,7 +22,7 @@ const options = {
     },
     servers: [
       {
-        url: "http://localhost:4000", // 요청 URL
+        url: `${sever_ip}`, // 요청 URL
       },
     ],
   },
@@ -43,75 +46,10 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: true })); // 폼 데이터를 파싱하기 위해 존재함.
 app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(specs));
 
-// 회원가입
-app.post("/register", async (req, res) => {
-  const { user_id, nickname, password, name, birth, phonenum, email } = req.body;
-
-  try {
-    const newUser = await prisma.user.create({
-      data: {
-        user_id, // Primary Key
-        nickname: nickname || null,
-        password, // 암호화 없이 저장
-        name: name || null,
-        birth: birth ? new Date(birth) : null, // Date로 변환
-        phonenum: phonenum || null,
-        email: email || null,
-      },
-    });
-
-    res.status(201).json({
-      message: "회원가입 성공",
-      user: {
-        user_id: newUser.user_id,
-        nickname: newUser.nickname,
-        email: newUser.email,
-      },
-    });
-  } catch (error) {
-    console.error("회원가입 에러:", error);
-    res.status(400).json({
-      message: "회원가입 실패",
-      error: error.message,
-    });
-  }
-});
-
-// 로그인
-app.post("/login", (req, res) => {
-  const { id, pw } = req.body;
-  const secretKey = process.env.JWT_SECRET;
-
-  const token = jwt.sign({ id, pw }, secretKey, {
-    expiresIn: "10h",
-  });
-
-  res.cookie("authToken", token, {
-    httpOnly: true,
-    secure: true,
-    sameSite: "Lax",
-    maxAge: 1000 * 60 * 60 * 10,
-  });
-
-  res.status(200).json({
-    message: "로그인 성공!",
-    token,
-  });
-});
-
-// 로그아웃
-app.get("/logout", (req, res) => {
-  res.clearCookie("authToken", {
-    httpOnly: true,
-    secure: true,
-    sameSite: "Lax",
-    path: "/",
-  });
-
-  res.status(200).json({
-    message: "로그아웃 성공! 쿠키가 삭제되었습니다.",
-  });
-});
+//유저관리
+app.post("/register", handleUserSignUp);
+app.post("/login", handleUserLogin);
+app.get("/logout", handleUserLogout);
 
 // 로그인 API
 /**
@@ -358,7 +296,6 @@ app.get("/logout", (req, res) => {
  *       404:
  *         description: 게시글이 없어요. 작성해주세요!
  */
-
 
 // 게시글 작성 API
 /**
