+++ test.http
@@ -1,13 +1,31 @@
-###회원가입
+###회원가입15.165.117.176
 POST http://localhost:4000/register
 Content-Type: application/json; charset=UTF-8
 
 {
-  "user_id": "1234",
-  "nickname": "chan",
-  "password": "1234",
-  "name": "hyck",
+  "user_id": "test3",
+  "nickname": "qqqq",
+  "password": "qqqq",
+  "name": "qqqq",
   "birth": "1990-01-01",
   "phonenum": "01067058085",
-  "email": "1234@naver.com"
+  "email": "qqqq@naver.com"
+}
+
+###로그인15.165.117.176
+POST http://localhost:4000/login
+Content-Type: application/json; charset=UTF-8
+
+{
+  "id": "1234",
+  "pw": "1234"
+}
+
+###
+POST https://w94yrmif2m.execute-api.ap-northeast-2.amazonaws.com/prod/login
+Content-Type: application/json; charset=UTF-8
+
+{
+  "id": "1",
+  "pw": "1234"
 }
